Secure views for logged in or superusers

+++ products/views.py
@@ -1,10 +1,12 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
 from .models import Product, Category
 from .forms import ProductForm
+
 
 def all_products(request):
     """ A view to display all products, including
@@ -72,8 +74,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Admin users can add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to do this')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -93,8 +100,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Admin users can edit products in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to do this')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -117,8 +129,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Admin users can delete products from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to do this')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request,'Product successfully deleted')
